+++ odbc-api/src/buffers/text_column.rs
@@ -1,11 +1,7 @@
-use crate::{
-    handles::{CData, Input},
-    DataType,
-};
+use crate::{DataType, handles::{CData, CDataMut, Input}};
 
-use super::{BindColArgs, ColumnBuffer};
 use log::debug;
-use odbc_sys::{CDataType, Len, Pointer, NULL_DATA};
+use odbc_sys::{CDataType, Len, NULL_DATA};
 use std::{cmp::min, convert::TryInto, ffi::c_void};
 
 /// A buffer intended to be bound to a column of a cursor. Elements of the buffer will contain a
@@ -22,17 +18,6 @@ pub struct TextColumn {
     /// with the same index. Please note that this value may be larger than `max_str_len` if the
     /// text has been truncated.
     indicators: Vec<Len>,
-}
-
-unsafe impl ColumnBuffer for TextColumn {
-    fn bind_arguments(&mut self) -> BindColArgs {
-        BindColArgs {
-            target_type: CDataType::Char,
-            target_value: self.values.as_mut_ptr() as Pointer,
-            target_length: (self.max_str_len + 1).try_into().unwrap(),
-            indicator: self.indicators.as_mut_ptr(),
-        }
-    }
 }
 
 impl TextColumn {
@@ -156,5 +141,15 @@ unsafe impl Input for TextColumn {
         DataType::Varchar {
             length: self.max_str_len.try_into().unwrap(),
         }
+    }
+}
+
+unsafe impl CDataMut for TextColumn {
+    fn mut_indicator_ptr(&mut self) -> *mut Len {
+        self.indicators.as_mut_ptr()
+    }
+
+    fn mut_value_ptr(&mut self) -> *mut c_void {
+        self.values.as_mut_ptr() as *mut c_void
     }
 }