+++ odbc-api/Changelog.md
@@ -8,6 +8,7 @@
 * `TextRowSet::new` renamed into `TextRowSet::for_cursor`.
 * trait `ParameterCollection` has a new method `parameter_set_size`.
 * `Parameter` trait is now inherting from `Input`.
+* `ColumnBuffer` trait is replaced by `CDataMut`.
 
 ## 0.9.1